==> 1_prelim_clean.ipynb <==
This notebook is a quick cleaning of the data just to get it usable for our baseline ML models. This is quick and dirty cleaning with further refinements (feature eng, cat encoding, etc) to be done later. Different models will require different cleaning so this mostly just combines all the seperate csv files and establishes train-val, and holdout_test set

import os, warnings, gc, requests, json, re
from requests.auth import HTTPDigestAuth
from pathlib import Path
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
%matplotlib inline
warnings.filterwarnings('ignore')
pd.set_option('display.max_columns', 50)

data_path = 'data/' # replace with name of your data file if different
key = '' # put your api key here
acc_id = '' # put your account id here

The cleaning function does some very simple cleaning like converting interest rates from strings to floats.

def cleaning(df):
            
    # drop non_loan columns
    index = df[df['loan_amnt'].isnull()].index
    df.drop(index=index, inplace=True)
    
    # convert % rate from str to float
    df['int_rate'] = df['int_rate'].str[:-1].astype(float)
    df['revol_util'] = df['revol_util'].str[:-1].astype(float)
    
    # convert term from str to int
    df['term'] = df['term'].str.strip().str[:2].astype(int)
    
    # drop weird old columns
    index = (df[(df['loan_status'] == 'Does not meet the credit policy. Status:Charged Off')
            |(df['loan_status'] == 'Does not meet the credit policy. Status:Fully Paid')]).index
    df.drop(index=index, inplace=True)
    
    # convert dates to useable formats
    df['earliest_cr_line'] = pd.to_datetime(df['earliest_cr_line'], format='%b-%Y')
    df['issue_d'] = pd.to_datetime(df['issue_d'], format='%b-%Y')
    
    # convert earliest cr_line from date to months since loan request
    df['earliest_cr_line'] = ((df['issue_d'] - df['earliest_cr_line']) / np.timedelta64(1, 'M')).astype(int)
    
    return df 

The historical data is split into different csv files. The get data function will clean each one, then combine them all into one dataframe.

def get_data(approved_files, data_path):
    df = pd.DataFrame()
    for file in approved_files:
        print(f'reading in {file}')
        temp_df = pd.read_csv(data_path+file,header=1)
        temp_df = cleaning(temp_df)
        df = pd.concat([df,temp_df],ignore_index=True)
    del temp_df
    return df

My data folder also contains info on rejected loans. Bc of this, make a list of files containing only the csv files containing approved and issued loans

files = os.listdir(data_path)
# only files with approved loans start with 'L'
approved_files = [f for f in files if f[0]=='L']

Clean and combine all historical loan files into one.

df = get_data(approved_files, data_path)

df.head()

df.shape

# to view all column names
#list(df.columns)

The dataframe currently holds information not available at the time of investment.

Enter your LC account info to preview the currently listed loans. From the currently available loans create a list of all available features. These are the features we know we can use in our ML models. Then try to match these features to the ones available in the Lending Club historical data.

api_key =  {'Authorization': key} 
acc_id = acc_id 

# get loan listings data
loans = 'https://api.lendingclub.com/api/investor/v1/loans/listing'
res = requests.get(loans, headers=api_key)
data = json.loads(res.text)

# grabs the available features
avail_cols = list(data['loans'][0].keys())

# to view columns from the listings api
#avail_cols

Feature names from the api differ from the historical data. Since it's mostly slight formatting differences regex functions will be used to match column names. A dictionary is created to map other feature names from the api to the historical data. Some of the features mapped in the dictionary could also be matched using regex

to_map = {'secAppCollections12MthsExMed': 'sec_app_collections_12_mths_ex_med',
          'secAppInqLast6Mths': 'sec_app_inq_last_6mths',
          'numAcctsEver120Ppd': 'num_accts_ever_120_pd',
          'inqLast6Mths': 'inq_last_6mths',
          'numTl120dpd2m': 'num_tl_120dpd_2m',
          'numTl30dpd': 'num_tl_30dpd',
          'numTl90gDpd24m': 'num_tl_90g_dpd_24m',
          'numTlOpPast12m': 'num_tl_op_past_12m',
          'collections12MthsExMed': 'collections_12_mths_ex_med',
          'isIncV': 'verification_status',
          'isIncVJoint': 'verification_status_joint',
          'openIl12m': 'open_il_12m',
          'openIl24m': 'open_il_24m',
          'openRv12m': 'open_rv_12m',
          'openRv24m': 'open_rv_24m',
          'secAppChargeoffWithin12Mths': 'sec_app_chargeoff_within_12_mths',
          'addrZip': 'zip_code',
          'accOpenPast24Mths': 'acc_open_past_24mths',
          'chargeoffWithin12Mths': 'chargeoff_within_12_mths',
          'inqLast12m': 'inq_last_12m',
          'delinq2Yrs': 'delinq_2yrs',
          'percentBcGt75': 'percent_bc_gt_75',
          'loanAmount': 'loan_amnt',
          'iLUtil': 'il_util',          
         }

# cols w info regarding loan performace from LC dataset
# this will get stored with the data used for modeling so we can see returns on investment later
cols_of_interest = ['issue_d','loan_status','total_pymnt', 'total_rec_int',
                    'total_rec_late_fee','total_rec_prncp', 'recoveries',
                    'collection_recovery_fee', 'last_pymnt_d']
                    

# cols dropped from the listed loans features (these features are not in historical data)
to_drop = ['reviewStatus', 'housingPayment', 'creditPullD', 'ilsExpD', 'mtgPayment', 'expD', 'acceptD',
          'investorCount','serviceFeeRate', 'disbursementMethod', 'listD', 'expDefaultRate',
          'reviewStatusD','fundedAmount']

The next block calls the lendingclub api to get a list of available loans. From this we get the available features to use for our models. You can use your api key and investor id.

# performs feature matching between features from api call and features from dataset
# some features were easy to match with regex but for others is was quick to manually write mapping
# this can probably be clean up further
api_cols = []
for col in avail_cols:
    if col in list(to_map.keys()):
        new_col = to_map[col]
        api_cols.append(new_col)
        continue
    if col in to_drop:
        continue
    new_col = re.sub(r'([A-Z])', r'_\1', col).lower()
    new_col = re.sub(r'([0-9])+', r'_\1', new_col).lower()
    api_cols.append(new_col)

api_cols[:10]

len(api_cols)

api_cols is a list of usable features from the api call listed the feature names in the historical data. Now reassign the dataframe into a df with only the features we want.

df = df[api_cols+cols_of_interest]

df.head()

# save data
df.to_pickle('data/prelim-all.pkl')

Next create train/val/test splits. Since only historical data will be avaible when choosing which loans to invest in, I chose to split the data by date. I only use 36m term loans instead of both 36 and 60m because I do not want my money locked in an investment for 5 years. 

Most recent data is Q3 2019 so 36m loans from Q3 2016 and earlier should be fully paid off or confirmed default by now. This most recent loan data from Q4 2015- Q3 2016 will be used as the test set.

Train and Validation sets will have loans from Q3 2015 and earlier

df = pd.read_pickle('data/prelim-all.pkl')

Chose to restric to 36 month loans because so we have full payment info on more recent data. The other loan term is 5 years and would limit the time periods available for a train/val/holdout data split.

# restrict to 36m loans
df = df[df['term']==36]

df.shape

Next a few features are added to later calculate annualized returns.

# how long until the loan was paid off/reached default status
df['loan_length'] = (pd.to_datetime(df['last_pymnt_d'])-df['issue_d']).dt.days

# loans that did not recieve any payments had NA values
# by adding 30 days it provides a timeframe to calculate returns
# NA values messed up the calculation
df['loan_length'] = df['loan_length'].fillna(30)

# loans that were paid off in the same month had a loan length of 0
# loan length of zero pushed returns to infinity
# loans also take some time to originate after initial investment
# adding 30 days creates cleaner and more realistic return calculations
df.loc[df['loan_length']==0, 'loan_length'] = 30

# recoveries are not factored into the last_pyment_d feature
# without this feature a loan can make the first payment, then default and make partial recovry pymnt
# and return on investment would be above 100%
# adding a year for the recovery peroid is also fairly realistic
df.loc[df['recoveries']>0, 'loan_length'] =df.loc[df['recoveries']>0, 'loan_length']+365

# standard annualized return formula
df['returns'] = (df['total_pymnt']/df['loan_amnt'])**(365/(df['loan_length'])) - 1


df.sort_values('returns').head()

df.sort_values('returns').tail()

The train/val/holdout split is designed to simulate future investments.
Train and validate on historical data, and the (test)holdout set is the most recent loan data with full payment/chargoff information (3 years ago).

# split train and test data
test_end = pd.datetime(2016,9,1)
test_start = pd.datetime(2015,10,1)
test_df = df[(df['issue_d'] >= test_start) & (df['issue_d'] <= test_end)]

# verify start and end dates for test data
test_df.issue_d.min(), test_df.issue_d.max()

#export
train_end = pd.datetime(2015,9,1)
train_df = df[df['issue_d'] <= train_end]

# verify start and end dates for test data
train_df.issue_d.min(), train_df.issue_d.max()

Verify no test loans leak into train data. Testing intersection between sets much faster than finding overlap between lists of ids.

train_ids = set(train_df.id)
test_ids = set(test_df.id)

train_ids.intersection(test_ids)

Check loan status within each data split.

train_df['loan_status'].value_counts()

train_df = train_df[(train_df['loan_status']=='Fully Paid')|(train_df['loan_status']=='Charged Off')]

train_df['loan_status'].value_counts()

test_df['loan_status'].value_counts()

# for simplicity
test_df = test_df[(test_df['loan_status']=='Fully Paid')|(test_df['loan_status']=='Charged Off')]

test_df['loan_status'].value_counts()

train_df['application_type'].value_counts()

test_df['application_type'].value_counts()

test_df = test_df[test_df['application_type']!='Joint App']

test_df['application_type'].value_counts()

train_df.to_pickle('data/api_train_df.pkl')
test_df.to_pickle('data/api_test_df.pkl')
