==> 2_rate_dollar_adjustments.ipynb <==
This notebook will adjust all dollar amounts by inflation rates, and convert loan interest rates to be relative to the prime rate at the time of loan issuance.

import warnings
warnings.filterwarnings('ignore')
import numpy as np
import pandas as pd
from pathlib import Path
pd.set_option('display.max_columns', 50)

data_path = 'data/'

train_df = pd.read_pickle(data_path+'api_train_df.pkl')
test_df = pd.read_pickle(data_path+'api_test_df.pkl')

To help with generalizing over time, convert income and payment amounts to inflation adjusted amounts. Also to generalize to different interest rate envornments over time, convert interest rates to interest relative to the prime rate

Prime rates and inflation rates downloaded from st louis Fed.

prime rate data: https://fred.stlouisfed.org/series/MPRIME
inflation rates: https://fred.stlouisfed.org/series/CPIAUCSL

# load prime rate
prime_rate = pd.read_csv(data_path+'MPRIME.csv')
prime_rate['DATE'] = pd.to_datetime(prime_rate['DATE'])

train_df = train_df.merge(prime_rate, left_on='issue_d', right_on='DATE')
test_df = test_df.merge(prime_rate, left_on='issue_d', right_on='DATE')

train_df['issue_m'] = train_df['issue_d'].dt.month
test_df['issue_m'] = test_df['issue_d'].dt.month

train_df['int_rate'] = train_df['int_rate'] - train_df['MPRIME']
test_df['int_rate'] = test_df['int_rate'] - test_df['MPRIME']


to_adj = ['annual_inc','loan_amnt','installment',
          'revol_bal', 'avg_cur_bal','bc_open_to_buy',
          'total_il_high_credit_limit','total_bc_limit',
          'total_rev_hi_lim','tot_hi_cred_lim','total_bal_ex_mort',
          'tot_cur_bal','tot_coll_amt','delinq_amnt']

# load inflation info
inflation = pd.read_csv(data_path+'CPIAUCSL.csv')
inflation['DATE'] = pd.to_datetime(inflation['DATE'])

# use prev month inflation because current month inflation will not be available at time of investment
# number provided is CPI index benchmarked to past date
inflation['prev_mo_cpi'] = inflation['CPIAUCSL'].shift()

train_df = train_df.merge(inflation, left_on='issue_d', right_on='DATE')
test_df = test_df.merge(inflation, left_on='issue_d', right_on='DATE')

inflation_adj_2016 = inflation[inflation['DATE']=='2016-01-01']['CPIAUCSL'].values[0]

inflation_adj_2016

# adjust all dollar amounts to be relative to jan 2016
for val in to_adj:
    train_df[val] = train_df[val]/train_df['prev_mo_cpi']
    train_df[val] = train_df[val]*inflation_adj_2016
    train_df[val] = train_df[val].round(decimals=2)
    
    test_df[val] = test_df[val]/test_df['prev_mo_cpi']
    test_df[val] = test_df[val]*inflation_adj_2016
    test_df[val] = test_df[val].round(decimals=2)

train_df.head()

train_df.drop(columns=['DATE_x','DATE_y'], inplace=True)
test_df.drop(columns=['DATE_x','DATE_y'], inplace=True)

train_df.to_pickle(data_path+'train_adj.pkl')
test_df.to_pickle(data_path+'test_adj.pkl')
