==> 3_data_cleaning_LGBM.ipynb <==
This notebook will clean data to use for LightGBM models.

import warnings
warnings.filterwarnings('ignore')
import numpy as np
import pandas as pd
from pathlib import Path
import pickle
pd.set_option('display.max_columns', 50)

data_path = 'data/'

train_df = pd.read_pickle(data_path+'train_adj.pkl').iloc[:,2:] # first two columns are id info
test_df = pd.read_pickle(data_path+'test_adj.pkl').iloc[:,2:]

train_df.head()

train_df.shape

Create list of features by removing loan performace info

to_drop = ['issue_d','desc','loan_status', 'total_pymnt','total_rec_int',
           'total_rec_late_fee','total_rec_prncp', 'recoveries', 
           'collection_recovery_fee', 'last_pymnt_d', 
           'loan_length', 'returns', 'CPIAUCSL']


x_cols = [x for x in list(train_df.columns) if x not in to_drop]

# emp length should be ordered
emp_len_map = {'< 1 year':0, '1 year':1, '2 years': 2, '3 years': 3,
               '4 years':4, '5 years':5, '6 years':6, '7 years':7,
               '8 years':8, '9 years':9, '10+ years': 10}

train_df['emp_length'] = train_df['emp_length'].map(emp_len_map)
test_df['emp_length'] = test_df['emp_length'].map(emp_len_map)

To encode the categorical/str columns to numbers:
1. Create a dictionary with encoding info for each column
2. For each column, with the training data create a dict mapping each category to an int
3. Apply the categorical mapping from str to ints on the train, val, and test data

Sklearn label encoder was not used because it has issues if there are values in the val set we are mapping that were not seen in the train encodings. We do not want to fit a label encoder on both train and val data because we are simulating training on historical data to validate against unseen data.

cat_encs = {}

for col in x_cols:
    if train_df[col].dtype == object:
        unique = train_df[col].unique()
        try: # keeps alphabetical order, good for grade and subgrade
            unique.sort()
            cat_encs[col] = dict(zip(unique, range(len(train_df[col].unique()))))
        except:
            cat_encs[col] = dict(zip(unique, range(len(train_df[col].unique()))))
        
for col in cat_encs.keys():
    train_df[col] = train_df[col].map(cat_encs[col], na_action='UNK')
    test_df[col] = test_df[col].map(cat_encs[col], na_action='UNK')

Some columns are missing all info, so remove those columns

drop_missing = []
for col in x_cols:
    if train_df[col].isnull().mean()>=0.99:
        drop_missing.append(col)

x_cols = [x for x in x_cols if x not in drop_missing]

returns_info = ['issue_d','desc','loan_status', 'total_pymnt','total_rec_int',
                'total_rec_late_fee','total_rec_prncp', 'recoveries', 
                'collection_recovery_fee', 'last_pymnt_d', 
                'loan_length', 'returns']

# these are the same for every loan
to_drop = ['term','application_type','verification_status_joint',
           'sec_app_earliest_cr_line']

# save data
train_df[x_cols+returns_info].to_pickle(data_path+'train_lgbm.pkl')
test_df[x_cols+returns_info].to_pickle(data_path+'test_lgbm.pkl')

cat_encs['emp_length']= emp_len_map

# Store cat encs
with open('cat_encs.pickle', 'wb') as f:
    pickle.dump(cat_encs, f)
